modal-outsourced
Sprawdź 2 razy czy to czegoś nie spieprzy.

Wyrzuciłem obsługę okienka modalnego do zewnętrznego pliku, którego działanie importuję do movieContainer.js
